Fix cyclic dependencies (hopefully)
