+++ shared/src/models/vehicle.ts
@@ -3,8 +3,7 @@ import { IsNumber, IsString, IsUUID, ValidateNested } from 'class-validator';
 import { uuid, uuidValidationOptions, UUID, UUIDSet } from '../utils';
 import { IsUUIDSet, IsValue } from '../utils/validators';
 import { IsPosition } from '../utils/validators/is-position';
-import type { WithPosition } from './utils';
-import { currentVehicleIdOf, getCreate, isInVehicle } from './utils';
+import { getCreate } from './utils';
 import { ImageProperties } from './utils/image-properties';
 import { Position } from './utils/position/position';
 
@@ -64,11 +63,4 @@ export class Vehicle {
     }
 
     static readonly create = getCreate(this);
-
-    static isInVehicle(vehicle: Vehicle, withPosition: WithPosition): boolean {
-        return (
-            isInVehicle(withPosition) &&
-            currentVehicleIdOf(withPosition) === vehicle.id
-        );
-    }
 }
