+++ shared/src/simulation/activities/delay-event.ts
@@ -4,7 +4,7 @@ import { getCreate } from '../../models/utils';
 import { UUID, uuidValidationOptions } from '../../utils';
 import { IsValue } from '../../utils/validators';
 import { ExerciseSimulationEvent, simulationEventTypeOptions } from '../events';
-import { sendSimulationEvent } from '../utils/simulated-region';
+import { sendSimulationEvent } from '../events/utils';
 import type {
     SimulationActivity,
     SimulationActivityState,
