+++ shared/src/simulation/index.ts
@@ -1,4 +1,3 @@
 export * from './activities';
 export * from './behaviors';
 export * from './events';
-export * from './utils';
