--- shared/src/simulation/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './simulated-region';
-export * from './vehicle';
-export * from './randomness';
-export * from './treatment';
