+++ shared/src/simulation/utils/simulated-region.ts
@@ -1,97 +1,16 @@
 import type { SimulatedRegion } from '../../models';
-import type { ExerciseState } from '../../state';
-import type { Mutable, UUID } from '../../utils';
-import { cloneDeepMutable } from '../../utils';
-import type { ExerciseSimulationActivityState } from '../activities';
-import { simulationActivityDictionary } from '../activities';
-import { simulationBehaviorDictionary } from '../behaviors';
-import type { ExerciseSimulationEvent } from '../events';
-import { TickEvent } from '../events/tick';
+import type { WithPosition } from '../../models/utils';
+import {
+    currentSimulatedRegionIdOf,
+    isInSimulatedRegion,
+} from '../../models/utils';
 
-export function simulateAllRegions(
-    draftState: Mutable<ExerciseState>,
-    tickInterval: number
-) {
-    Object.values(draftState.simulatedRegions).forEach((simulatedRegion) => {
-        simulateSingleRegion(draftState, simulatedRegion, tickInterval);
-    });
-}
-
-function simulateSingleRegion(
-    draftState: Mutable<ExerciseState>,
-    simulatedRegion: Mutable<SimulatedRegion>,
-    tickInterval: number
-) {
-    sendSimulationEvent(simulatedRegion, TickEvent.create(tickInterval));
-    handleSimulationEvents(draftState, simulatedRegion);
-    tickActivities(draftState, simulatedRegion, tickInterval);
-}
-
-function tickActivities(
-    draftState: Mutable<ExerciseState>,
-    simulatedRegion: Mutable<SimulatedRegion>,
-    tickInterval: number
-) {
-    Object.values(simulatedRegion.activities).forEach((activityState) => {
-        simulationActivityDictionary[activityState.type].tick(
-            draftState,
-            simulatedRegion,
-            activityState as any,
-            tickInterval,
-            () => {
-                terminateActivity(
-                    draftState,
-                    simulatedRegion,
-                    activityState.id
-                );
-            }
-        );
-    });
-}
-
-function handleSimulationEvents(
-    draftState: Mutable<ExerciseState>,
-    simulatedRegion: Mutable<SimulatedRegion>
-) {
-    simulatedRegion.behaviors.forEach((behaviorState) => {
-        simulatedRegion.inEvents.forEach((event) => {
-            simulationBehaviorDictionary[behaviorState.type].handleEvent(
-                draftState,
-                simulatedRegion,
-                behaviorState as any,
-                event
-            );
-        });
-    });
-    simulatedRegion.inEvents = [];
-}
-
-export function sendSimulationEvent(
-    simulatedRegion: Mutable<SimulatedRegion>,
-    event: ExerciseSimulationEvent
-) {
-    simulatedRegion.inEvents.push(cloneDeepMutable(event));
-}
-
-export function addActivity(
-    simulatedRegion: Mutable<SimulatedRegion>,
-    activityState: ExerciseSimulationActivityState
-) {
-    simulatedRegion.activities[activityState.id] =
-        cloneDeepMutable(activityState);
-}
-
-export function terminateActivity(
-    draftState: Mutable<ExerciseState>,
-    simulatedRegion: Mutable<SimulatedRegion>,
-    activityId: UUID
-) {
-    const activityType = simulatedRegion.activities[activityId]?.type;
-    if (activityType) {
-        const activity = simulationActivityDictionary[activityType];
-        if (activity.onTerminate) {
-            activity.onTerminate(draftState, simulatedRegion, activityId);
-        }
-        delete simulatedRegion.activities[activityId];
-    }
+export function isInThisSimulatedRegion(
+    region: SimulatedRegion,
+    withPosition: WithPosition
+): boolean {
+    return (
+        isInSimulatedRegion(withPosition) &&
+        currentSimulatedRegionIdOf(withPosition) === region.id
+    );
 }
